backend/requirements.txt
flask==2.3.3
flask-sqlalchemy==3.0.5
flask-cors==4.0.0
python-dotenv==1.0.0
pyjwt==2.8.0
selenium==4.15.0
requests==2.31.0
mercadopago==2.0.1
